refactor(web): extract useChatCore to unify homepage and wiki chat logic
- New useChatCore hook: shared SSE subscription, event processing, message
  state management, multi-turn, cancel, submitToolResult, and reset
- useChat delegates to core, retains only DB persistence + loadConversation
- useWikiChat delegates to core, retains only wiki API dispatch + startOperation
- WikiChatPanel now reuses shared ChatComposer component
- Removed ~650 lines of duplicated code
